+++ src/pages/Validators/Stats/AvtiveEra.tsx
@@ -5,7 +5,7 @@ import moment from 'moment';
 import { useNetworkMetrics } from '../../../contexts/Network';
 import { useSessionEra } from '../../../contexts/SessionEra';
 import { useEraTimeLeft } from '../../../library/Hooks/useEraTimeLeft';
-import { Pie } from '../Pie';
+import { Pie } from '../../../library/StatBoxList/Pie';
 
 const ActiveEraStatBox = () => {
   const { metrics } = useNetworkMetrics();
